Make action unit tab generic

+++ src/main/java/fr/siamois/infrastructure/database/repositories/actionunit/ActionUnitRepository.java
@@ -28,7 +28,7 @@ public interface ActionUnitRepository extends CrudRepository<ActionUnit, Long> {
             value = "SELECT COUNT(*) FROM action_unit_spatial_context WHERE fk_spatial_unit_id = :spatialUnitId",
             nativeQuery = true
     )
-    long countBySpatialContext(@Param("spatialUnitId") Long spatialUnitId);
+    Integer countBySpatialContext(@Param("spatialUnitId") Long spatialUnitId);
 
     long countByCreatedByInstitution(Institution institution);
